Time: 20 ms (68.74%), Space: 70.4 MB (80.91%) - LeetHub

+++ 1482-minimum-number-of-days-to-make-m-bouquets/1482-minimum-number-of-days-to-make-m-bouquets.cpp
@@ -1,5 +1,5 @@
 class Solution {
-    bool isPossible(vector<int>&bloomDay, long long m, long long k, int mid)
+    bool isPossible(vector<int>&bloomDay, long long m, long long k, long long mid)
     {   
         int count =0, noofbouquet=0;
        for(int i = 0; i<bloomDay.size(); i++)
